Moved common defs from gpuMILibrary to gpuKnnLibrary.
In preparation for gpuCMILibrary.
